adjusted code to formatters/linters
